test: migrate to bdd test framework
Use bdd framework can gain better management for subprocess resources.
